Separate LayerList and LayerListWin
This splits the LayerList component into two separate components:
LayerList (the list itself)ynced with the map and the LayerListWin
(only a window holding the LayerList).
